Rename pipeline_endponit to smooch_endpoint
