+++ scripts/deploy_pipeline.js
@@ -1,20 +1,20 @@
 'use strict';
 
-const pipeline = require('../pipeline_endpoint');
+const smooch = require('../smooch_endpoint');
 const config = require('../pipeline_config');
 
 async function removeOldProcessors() {
-	const processorIds = (await pipeline.listProcessors()).map((processor) => processor._id);
+	const processorIds = (await smooch.listProcessors()).map((processor) => processor._id);
 	for (const processorId of processorIds) {
 		console.log('delete processor', processorId);
-		await pipeline.deleteProcessor(processorId);
+		await smooch.deleteProcessor(processorId);
 	}
 }
 
 async function addNewProcessors() {
 	const ids = [];
 	for (const endpoint of config.orderedEndpoints) {
-		const data = await pipeline.createProcessor(endpoint);
+		const data = await smooch.createProcessor(endpoint);
 		console.log('add new processors', data);
 		ids.push(data.id);
 	}
@@ -23,10 +23,10 @@ async function addNewProcessors() {
 }
 
 async function deployPipeline() {
-	console.log('deploying pipeline to Smooch');
+	console.log('deploying smooch to Smooch');
 	await removeOldProcessors();
 	const processorIds = await addNewProcessors();
-	await pipeline.setPipeline(processorIds);
+	await smooch.setPipeline(processorIds);
 	console.log('deployment complete');
 }
